memoryos-playground/requirements.txt
# MemoryOS Core Dependencies
# Core scientific computing and ML libraries
numpy==1.24.*
sentence-transformers>=2.7.0,<3.0.0 # Updated for Qwen model support
transformers>=4.51.0              # Required for newer sentence-transformer features
FlagEmbedding>=1.2.9                # For BGE-M3 model support

faiss-gpu>=1.7.0,<2.0.0

openai
# Web framework (for demo)
flask>=2.0.0,<3.0.0

# Optional utilities
python-dotenv>=0.19.0,<2.0.0

# Development and testing (optional)
# pytest>=7.0.0,<8.0.0
# pytest-asyncio>=0.20.0,<1.0.0

# Additional dependencies for compatibility
typing-extensions>=4.0.0,<5.0.0
regex>=2022.1.18
